fixed one more warning

+++ DebugView++Lib/ProcessMonitor.cpp
@@ -57,7 +57,7 @@ void ProcessMonitor::Run()
 		handles[0] = m_event.get();
 		size_t processCount = m_processes.size();
 		auto count = std::min<size_t>(processCount, MAXIMUM_WAIT_OBJECTS - 1);
-		for (int i = 0; i < count; ++i)
+		for (size_t i = 0; i < count; ++i)
 			handles[i + 1] = m_processes[(offset + i) % processCount].handle;
 		DWORD timeout = static_cast<size_t>(count) < m_processes.size() ? 1000 : INFINITE;
 		auto result = Win32::WaitForAnyObject(handles.data(), handles.data() + count + 1, timeout);
